# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for EVE Online trading, manufacturing and profit optimization\n\nFeatures:\n- Intra-region trading route calculation\n- Character skills and ship fitting integration\n- Deterministic cargo/warp speed calculations\n- Market data caching with staleness tracking\n- EVE SSO authentication",
        "title": "EVE-O-Provit API",
        "contact": {
            "name": "EVE-O-Provit Team",
            "url": "https://github.com/Sternrassler/eve-o-provit",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/characters/{characterId}/fitting/{shipTypeId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve character's ship fitting with deterministic bonus calculations\nCalculates effective cargo capacity, warp speed with skills and modules\nUses EVE dogma engine with stacking penalties",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fitting"
                ],
                "summary": "Get character ship fitting",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 12345678,
                        "description": "Character ID",
                        "name": "characterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 650,
                        "description": "Ship Type ID",
                        "name": "shipTypeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Force cache refresh",
                        "name": "refresh",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CharacterFittingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/characters/{characterId}/skills": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve character skills from ESI with Redis caching\nGraceful degradation: Returns default skills (level 0) if ESI fails",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Character"
                ],
                "summary": "Get character skills",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 12345678,
                        "description": "Character ID",
                        "name": "characterId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CharacterSkillsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "CharacterFittingResponse": {
            "type": "object",
            "properties": {
                "applied_skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Gallente Hauler V",
                        "Navigation V"
                    ]
                },
                "base_cargo_hold_m3": {
                    "type": "number",
                    "example": 2700
                },
                "base_fuel_bay_m3": {
                    "type": "number",
                    "example": 0
                },
                "base_warp_speed_au_s": {
                    "type": "number",
                    "example": 3
                },
                "cached_until": {
                    "type": "string",
                    "example": "2025-11-12T11:00:00Z"
                },
                "character_id": {
                    "type": "integer",
                    "example": 12345678
                },
                "effective_cargo_m3": {
                    "type": "number",
                    "example": 9656.9
                },
                "fitted_modules": {
                    "type": "integer",
                    "example": 5
                },
                "inertia_modifier": {
                    "type": "number",
                    "example": 1
                },
                "ship_name": {
                    "type": "string",
                    "example": "Nereus"
                },
                "ship_type_id": {
                    "type": "integer",
                    "example": 650
                },
                "warp_speed_au_s": {
                    "type": "number",
                    "example": 6.87
                }
            }
        },
        "CharacterSkill": {
            "type": "object",
            "properties": {
                "active_skill_level": {
                    "type": "integer",
                    "example": 5
                },
                "skill_id": {
                    "type": "integer",
                    "example": 3340
                },
                "skillpoints_in_skill": {
                    "type": "integer",
                    "example": 256000
                },
                "trained_skill_level": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "CharacterSkillsResponse": {
            "type": "object",
            "properties": {
                "cached_until": {
                    "type": "string",
                    "example": "2025-11-12T11:00:00Z"
                },
                "character_id": {
                    "type": "integer",
                    "example": 12345678
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CharacterSkill"
                    }
                },
                "total_sp": {
                    "type": "integer",
                    "example": 50000000
                },
                "unallocated_sp": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "message": {
                    "type": "string",
                    "example": "Detailed error message"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "EVE SSO Bearer Token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}